ux(variants): hide variants entity
