+++ websites/frontend/app/modules/strategy/types.ts
@@ -3,8 +3,6 @@ import { Rule, RuleUpdateDto } from "../rules/types";
 export enum ValueToServe {
   Boolean = "Boolean",
   Variant = "Variant",
-  String = "String",
-  Number = "Number",
 }
 
 export enum TargetEntity {
