--- websites/frontend/app/routes/dashboard.projects.$id.environments.$env.flags.$flagId.variants._index.tsx
@@ -1,156 +1,0 @@
-import {
-  ActionFunction,
-  LoaderFunction,
-  V2_MetaFunction,
-} from "@remix-run/node";
-import { useLoaderData, useSearchParams } from "@remix-run/react";
-import { SuccessBox } from "~/components/Boxes/SuccessBox";
-import { CreateButton } from "~/components/Buttons/CreateButton";
-import { Card, CardContent } from "~/components/Card";
-import { EmptyState } from "~/components/EmptyState";
-import { PageTitle } from "~/components/PageTitle";
-import { DashboardLayout } from "~/layouts/DashboardLayout";
-import { useEnvironment } from "~/modules/environments/contexts/useEnvironment";
-import { getEnvMetaTitle } from "~/modules/environments/services/getEnvMetaTitle";
-import { FlagEnvMenu } from "~/modules/flags/components/FlagEnvMenu";
-import { useFlagEnv } from "~/modules/flags/contexts/useFlagEnv";
-import { toggleFlagAction } from "~/modules/flags/form-actions/toggleFlagAction";
-import { getFlagEnvMetaTitle } from "~/modules/flags/services/getFlagEnvMetaTitle";
-import { useProject } from "~/modules/projects/contexts/useProject";
-import { getProjectMetaTitle } from "~/modules/projects/services/getProjectMetaTitle";
-import { VariantList } from "~/modules/variants/components/VariantList";
-import { getVariants } from "~/modules/variants/services/getVariants";
-import { Variant } from "~/modules/variants/types";
-import { getSession } from "~/sessions";
-
-export const meta: V2_MetaFunction = ({ matches, params }) => {
-  const projectName = getProjectMetaTitle(matches);
-  const envName = getEnvMetaTitle(matches, params.env!);
-  const flagName = getFlagEnvMetaTitle(matches);
-
-  return [
-    {
-      title: `Progressively | ${projectName} | ${envName} | Flags | ${flagName} | Variants`,
-    },
-  ];
-};
-
-interface LoaderData {
-  variants: Array<Variant>;
-}
-
-export const loader: LoaderFunction = async ({
-  request,
-  params,
-}): Promise<LoaderData> => {
-  const session = await getSession(request.headers.get("Cookie"));
-  const authCookie = session.get("auth-cookie");
-
-  const variants: Array<Variant> = await getVariants(
-    params.env!,
-    params.flagId!,
-    authCookie
-  );
-
-  return {
-    variants,
-  };
-};
-
-type ActionDataType = null | {
-  errors?: { [key: string]: string | undefined };
-};
-
-export const action: ActionFunction = async ({
-  request,
-  params,
-}): Promise<ActionDataType> => {
-  const session = await getSession(request.headers.get("Cookie"));
-  const authCookie = session.get("auth-cookie");
-  const formData = await request.formData();
-  const type = formData.get("_type");
-
-  if (type === "toggle-flag") {
-    return toggleFlagAction(formData, params, authCookie);
-  }
-
-  return null;
-};
-
-export default function Variants() {
-  const { project } = useProject();
-  const { environment } = useEnvironment();
-  const { flagEnv } = useFlagEnv();
-  const { variants } = useLoaderData<LoaderData>();
-  const [searchParams] = useSearchParams();
-  const currentFlag = flagEnv.flag;
-
-  const isVariantRemoved = searchParams.get("variantRemoved") || undefined;
-  const isVariantAdded = searchParams.get("newVariant") || undefined;
-
-  const hasVariants = variants.length > 0;
-
-  return (
-    <DashboardLayout
-      subNav={
-        <FlagEnvMenu
-          projectId={project.uuid}
-          envId={environment.uuid}
-          flagEnv={flagEnv}
-        />
-      }
-      status={
-        isVariantRemoved ? (
-          <SuccessBox id="variant-removed">
-            The variant has been successfully removed.
-          </SuccessBox>
-        ) : isVariantAdded ? (
-          <SuccessBox id="variant-added">
-            The variant has been successfully created.
-          </SuccessBox>
-        ) : null
-      }
-    >
-      <PageTitle
-        value="Variants"
-        action={
-          hasVariants && (
-            <CreateButton
-              to={`/dashboard/projects/${project.uuid}/environments/${environment.uuid}/flags/${currentFlag.uuid}/variants/create`}
-            >
-              Create a variant
-            </CreateButton>
-          )
-        }
-      />
-
-      {!hasVariants && (
-        <Card>
-          <CardContent>
-            <EmptyState
-              titleAs="h2"
-              title="No variants found"
-              description={"There are no variants for this flag."}
-              action={
-                <CreateButton
-                  to={`/dashboard/projects/${project.uuid}/environments/${environment.uuid}/flags/${currentFlag.uuid}/variants/create`}
-                >
-                  Create a variant
-                </CreateButton>
-              }
-            />
-          </CardContent>
-        </Card>
-      )}
-
-      {hasVariants && (
-        <VariantList
-          variants={variants}
-          projectId={project.uuid}
-          envId={environment.uuid}
-          flagId={currentFlag.uuid}
-        />
-      )}
-    </DashboardLayout>
-  );
-}
