+++ websites/frontend/app/routes/dashboard.projects.$id.environments.$env.flags.$flagId.variants.create.tsx
@@ -47,7 +47,7 @@ export const action: ActionFunction = async ({
   }
 
   return redirect(
-    `/dashboard/projects/${params.id}/environments/${params.env}/flags/${params.flagId}/variants?newVariant=true#variant-added`
+    `/dashboard/projects/${params.id}/environments/${params.env}/flags/${params.flagId}?newVariant=true#variant-added`
   );
 };
