+++ websites/frontend/cypress/e2e/projects/id/environments/envId/flags/flagId/variants/create.spec.ts
@@ -58,7 +58,7 @@ describe("/dashboard/projects/[id]/environments/[envId]/flags/[flagId]/variants/
 
         cy.url().should(
           "include",
-          "/dashboard/projects/1/environments/1/flags/1/variants?newVariant=true"
+          "/dashboard/projects/1/environments/1/flags/1?newVariant=true"
         );
 
         cy.get(".success-box")
